Change header names, move some C functions around
